requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization libraries
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
folium>=0.14.0

# Machine learning libraries
scikit-learn>=1.3.0
xgboost>=1.7.0
tensorflow>=2.13.0
statsmodels>=0.14.0

# Web framework for dashboard
streamlit>=1.25.0

# Data processing and utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Geospatial libraries
geopandas>=0.13.0
shapely>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration
PyYAML>=6.0

# Progress bars and utilities
tqdm>=4.65.0